Reworked Guid test to verify UUID1 features instead of UUID4
