+++ lib/Doctrine/DBAL/Platforms/PostgreSqlPlatform.php
@@ -1019,7 +1019,7 @@ class PostgreSqlPlatform extends AbstractPlatform
      */
     public function getGuidExpression()
     {
-        return 'UUID_GENERATE_V4()';
+        return 'UUID_GENERATE_V1()';
     }
 
     /**
